requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
pymupdf
langchain-text-splitters
langchain-openai
chromadb
langchain_chroma
duckdb
pandas
openai
langchain-groq
pymupdf
mysql-connector-python
SQLAlchemy
validators==0.28.1
streamlit



